requirements.txt
streamlit==1.41.1
beautifulsoup4==4.12.3
selenium==4.27.1
webdriver-manager==4.0.2
transformers==4.47.1
accelerate==1.2.1
newspaper3k==0.2.8
lxml_html_clean==0.4.1
requests~=2.32.3